Reduce to core library
