+++ Router/Interfaces/ILink.cs
@@ -1,0 +1,10 @@
+using System;
+using QuikGraph;
+
+namespace Router.Interfaces
+{
+    public interface ILink : IEdge<IVertex>
+    {
+        Guid Id { get; }
+    }
+}
