--- Router/MainWindow.xaml
@@ -1,22 +1,0 @@
-﻿<Window x:Class="Router.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:m="http://materialdesigninxaml.net/winfx/xaml/themes"
-        xmlns:v="clr-namespace:Router.Views"
-        mc:Ignorable="d"
-        Title="Router" Height="600" Width="900"
-        TextElement.Foreground="{DynamicResource MaterialDesignBody}"
-        Background="{DynamicResource MaterialDesignPaper}"
-        TextElement.FontWeight="Medium"
-        TextElement.FontSize="14"
-        FontFamily="{m:MaterialDesignFont}">
-
-    <Grid>
-        <m:ColorZone Padding="5" 
-                     m:ElevationAssist.Elevation="Dp2">
-            <v:GraphView/>
-        </m:ColorZone>
-    </Grid>
-</Window>
