+++ Router/Router.csproj
@@ -1,29 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>net8.0-windows</TargetFramework>
-    <UseWPF>true</UseWPF>
+    <OutputType>Library</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="CommunityToolkit.Mvvm" Version="8.3.2" />
-    <PackageReference Include="MaterialDesignThemes" Version="4.9.0" />
-    <PackageReference Include="MathNet.Numerics" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.1" />
-    <PackageReference Include="ObjectTK" Version="2.0.0-pre.9" />
-    <PackageReference Include="OpenTK.GLWpfControl" Version="4.3.3" />
     <PackageReference Include="QuikGraph" Version="2.5.0" />
-    <PackageReference Include="System.Drawing.Common" Version="8.0.10" />
-  </ItemGroup>
-
-  <ItemGroup>
-	<Reference Include="GLGraphs">
-	  <HintPath>Libs\GLGraphs.dll</HintPath>
-	</Reference>
-    <Reference Include="GLGraphs.Wpf">
-      <HintPath>Libs\GLGraphs.Wpf.dll</HintPath>
-    </Reference>
   </ItemGroup>
 </Project>
