--- Router/Views/ConeView.xaml
@@ -1,46 +1,0 @@
-﻿<UserControl x:Class="Router.Views.ConeView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:c="clr-namespace:Router.Controls"
-             xmlns:m="clr-namespace:Router.Model"
-             xmlns:s="clr-namespace:System;assembly=mscorlib"
-             xmlns:e="clr-namespace:Router.Enums"
-             mc:Ignorable="d" 
-             d:DataContext="{x:Type m:Link}">
-    
-    <UserControl.Resources>
-        <ObjectDataProvider x:Key="FiberTypes" 
-                            MethodName="GetValues"
-                            ObjectType="{x:Type s:Enum}">
-            <ObjectDataProvider.MethodParameters>
-                <x:Type TypeName="e:FiberType"/>
-            </ObjectDataProvider.MethodParameters>
-        </ObjectDataProvider>
-    </UserControl.Resources>
-
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="auto" />
-            <RowDefinition Height="auto" />
-            <RowDefinition Height="auto" />
-        </Grid.RowDefinitions>
-
-        <GroupBox Style="{StaticResource MaterialDesignCardGroupBox}"
-                  Header="ID"
-                  Grid.Row="0">
-            <c:DefaultTextTextBox Style="{StaticResource CustomTextBox}" DefaultText="ID for a new link" Text="{Binding ID}" />
-        </GroupBox>
-        <GroupBox Style="{StaticResource MaterialDesignCardGroupBox}"
-                  Header="Length"
-                  Grid.Row="1">
-            <c:DefaultTextTextBox Style="{StaticResource CustomTextBox}" DefaultText="Length for a new link" Text="{Binding Weight}" />
-        </GroupBox>
-        <GroupBox Style="{StaticResource MaterialDesignCardGroupBox}"
-                  Header="Fiber type"
-                  Grid.Row="2">
-            <ComboBox Style="{StaticResource MaterialDesignComboBox}" ItemsSource="{Binding Source={StaticResource FiberTypes}}" SelectedItem="{Binding FiberType}" />
-        </GroupBox>
-    </Grid>
-</UserControl>
