--- Router/Views/ConeView.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace Router.Views
-{
-    /// <summary>
-    /// Interaction logic for ConeView.xaml
-    /// </summary>
-    public partial class ConeView : UserControl
-    {
-        public ConeView()
-        {
-            InitializeComponent();
-        }
-    }
-}
