--- Router/Views/GraphView.xaml
@@ -1,98 +1,0 @@
-﻿<UserControl x:Class="Router.Views.GraphView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:m="http://materialdesigninxaml.net/winfx/xaml/themes"
-             xmlns:e="clr-namespace:Router.Enums"
-             xmlns:c="clr-namespace:Router.Controls"
-             xmlns:vm="clr-namespace:Router.ViewModels"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-    <UserControl.DataContext>
-        <vm:GraphViewModel />
-    </UserControl.DataContext>
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="auto"/>
-            <RowDefinition Height="*" />
-            <RowDefinition Height="auto" />
-        </Grid.RowDefinitions>
-
-        <m:DrawerHost Grid.Row="0" 
-                      IsRightDrawerOpen="{Binding Mode, Converter={StaticResource GraphModeConverter}, ConverterParameter={x:Static e:GraphMode.Edit}}" 
-                      RightDrawerBackground="Transparent"
-                      OpenMode="Standard">
-            <m:DrawerHost.RightDrawerContent>
-                <StackPanel Orientation="Horizontal">
-                    <m:Chip Content="Nodes:"
-                            VerticalAlignment="Center"/>
-                    <RadioButton Style="{StaticResource TerminalMode}" 
-                                 GroupName="Nodes"
-                                 IsEnabled="{Binding Mode, Converter={StaticResource GraphModeConverter}, ConverterParameter={x:Static e:GraphMode.Edit}}" 
-                                 IsChecked="{Binding NodeMode, Converter={StaticResource NodeModeConverter}, ConverterParameter={x:Static e:NodeType.Terminal}}"/>
-                    <RadioButton Style="{StaticResource OLAMode}"
-                                 GroupName="Nodes"
-                                 IsEnabled="{Binding Mode, Converter={StaticResource GraphModeConverter}, ConverterParameter={x:Static e:GraphMode.Edit}}" 
-                                 IsChecked="{Binding NodeMode, Converter={StaticResource NodeModeConverter}, ConverterParameter={x:Static e:NodeType.OLA}}"/>
-                    <RadioButton Style="{StaticResource ROADMMode}"
-                                 GroupName="Nodes"
-                                 IsEnabled="{Binding Mode, Converter={StaticResource GraphModeConverter}, ConverterParameter={x:Static e:GraphMode.Edit}}" 
-                                 IsChecked="{Binding NodeMode, Converter={StaticResource NodeModeConverter}, ConverterParameter={x:Static e:NodeType.ROADM}}"/>
-
-                    <m:Chip Content="Spans:"
-                            VerticalAlignment="Center"/>
-                    <RadioButton Style="{StaticResource SimplexMode}"
-                                 GroupName="Links"
-                                 IsEnabled="{Binding Mode, Converter={StaticResource GraphModeConverter}, ConverterParameter={x:Static e:GraphMode.Edit}}" 
-                                 IsChecked="{Binding LinkMode, Converter={StaticResource LinkModeConverter}, ConverterParameter={x:Static e:LinkType.Simplex}}"/>
-                    <RadioButton Style="{StaticResource DuplexMode}" 
-                                 GroupName="Links"
-                                 IsEnabled="{Binding Mode, Converter={StaticResource GraphModeConverter}, ConverterParameter={x:Static e:GraphMode.Edit}}" 
-                                 IsChecked="{Binding LinkMode, Converter={StaticResource LinkModeConverter}, ConverterParameter={x:Static e:LinkType.Duplex}}"/>
-                </StackPanel>
-            </m:DrawerHost.RightDrawerContent>
-            <StackPanel Orientation="Horizontal">
-              
-                <RadioButton Style="{StaticResource SelectMode}"
-                             IsChecked="{Binding Mode, Converter={StaticResource GraphModeConverter}, ConverterParameter={x:Static e:GraphMode.Select}}"/>
-                <RadioButton Style="{StaticResource EditMode}"
-                             IsChecked="{Binding Mode, Converter={StaticResource GraphModeConverter}, ConverterParameter={x:Static e:GraphMode.Edit}}"/>
-            </StackPanel>
-        </m:DrawerHost>
-
-        <m:DrawerHost Grid.Row="1"
-                      IsRightDrawerOpen="{Binding InDrawerViewMode}" 
-                      RightDrawerBackground="Transparent"
-                      OpenMode="{Binding Mode, Converter={StaticResource OpenModeConverter}}">
-            <m:DrawerHost.RightDrawerContent>
-                <ContentPresenter Content="{Binding SelectedElement}"/>
-            </m:DrawerHost.RightDrawerContent>
-            <c:GraphAreaControl Height="500" 
-                                State="{Binding Network.State}"
-                                Settings="{Binding Network.Settings}" 
-                                VertexRequested="{Binding AddVertex}"/>
-
-        </m:DrawerHost>
-
-        <StackPanel Grid.Row="2"
-                    VerticalAlignment="Bottom">
-            <Separator />
-            <StatusBar Background="Beige" >
-                <StatusBarItem>
-                    <StackPanel Orientation="Horizontal">
-                        <TextBlock Text="Node count:    "/>
-                        <TextBlock Text="{Binding VertexCount}"/>
-                    </StackPanel>
-                </StatusBarItem>
-                <Separator />
-                <StatusBarItem>
-                    <StackPanel Orientation="Horizontal">
-                        <TextBlock Text="Link count:    "/>
-                        <TextBlock Text="{Binding Network.EdgeCount}"/>
-                    </StackPanel>
-                </StatusBarItem>
-            </StatusBar>
-        </StackPanel>
-    </Grid>
-</UserControl>
