--- Router/Views/GraphView.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace Router.Views
-{
-    /// <summary>
-    /// Interaction logic for GraphView.xaml
-    /// </summary>
-    public partial class GraphView : UserControl
-    {
-        public GraphView()
-        {
-            InitializeComponent();
-        }
-    }
-}
