--- Router/Views/NodeView.xaml
@@ -1,40 +1,0 @@
-﻿<UserControl x:Class="Router.Views.NodeView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:c="clr-namespace:Router.Controls"
-             xmlns:m="clr-namespace:Router.Model"
-             xmlns:s="clr-namespace:System;assembly=mscorlib"
-             xmlns:e="clr-namespace:Router.Enums"
-             mc:Ignorable="d" 
-             d:DataContext="{x:Type m:Node}">
-
-    <UserControl.Resources>
-        <ObjectDataProvider x:Key="NodeTypes" 
-                            MethodName="GetValues"
-                            ObjectType="{x:Type s:Enum}">
-            <ObjectDataProvider.MethodParameters>
-                <x:Type TypeName="e:NodeType"/>
-            </ObjectDataProvider.MethodParameters>
-        </ObjectDataProvider>
-    </UserControl.Resources>
-    
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="auto" />
-            <RowDefinition Height="auto" />
-        </Grid.RowDefinitions>
-
-        <GroupBox Style="{StaticResource MaterialDesignCardGroupBox}"
-                  Header="Name"
-                  Grid.Row="0">
-            <c:DefaultTextTextBox Style="{StaticResource CustomTextBox}" DefaultText="Name for a new node" Text="{Binding Name}" />
-        </GroupBox>
-        <GroupBox Style="{StaticResource MaterialDesignCardGroupBox}"
-                  Header="Type"
-                  Grid.Row="1">
-            <ComboBox Style="{StaticResource MaterialDesignComboBox}" ItemsSource="{Binding Source={StaticResource NodeTypes}}" SelectedItem="{Binding Type}" />
-        </GroupBox>
-    </Grid>
-</UserControl>
