--- Router/Views/NodeView.xaml.cs
@@ -1,16 +1,0 @@
-﻿using System.Windows.Controls;
-
-
-namespace Router.Views
-{
-    /// <summary>
-    /// Interaction logic for NodeView.xaml
-    /// </summary>
-    public partial class NodeView : UserControl
-    {
-        public NodeView()
-        {
-            InitializeComponent();
-        }
-    }
-}
